interfaces update, modularity enhanced
