+++ actions/stop_action.py
@@ -10,5 +10,4 @@ class StopAction(Action):
     def execute(self) -> None:
         # TODO check this logic because the function in the thread probably runs differently
         # and this will miss things up
-        self.dispatcher.serial_interface.terminate_serial_connection()
         self.dispatcher.serial_interface.terminate()
